Prevent fragmentation for timedummies for large T
Previously, when the time dimension was large, the line `df[name] = ...`
caused pandas to issue a performance warning that the dataframe got
highly fragmented.
This cluttered the console output, and potentially caused reduced
performance.

Now, the dummies are calculated using the `get_dummies`
method from pandas, and concatenated to the dataframe to all at once.
However, concatenating cannot be done in-place like inserting new
columns can, so the function signature had to be adjusted to return the
new dataframe.